+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Test cases when nothing changes in a deployment
 - Verbose logging of most kubectl commands
-- LogFilename parameter to Get-JobStatus, Get-PodStatus, and Write-PodLog
+- LogFileFolder parameter to Get-JobStatus, Get-PodStatus, and Write-PodLog
 - Added --wait to helm rollback
 
 ### Fixed